workflows/bonnke/brain/tasks/requirements.txt
langgraph==1.0.3
datetime==5.5
pydantic==2.12.4
langchain_openai==1.0.2
psycopg-binary==3.2.12
uuid==1.30
dotenv==0.9.9